refactor(slack): move webhook primitives into folder
