+++ packages/adapter-slack/tsup.config.ts
@@ -1,7 +1,10 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/webhook.ts"],
+  entry: {
+    index: "src/index.ts",
+    webhook: "src/webhook/index.ts",
+  },
   format: ["esm"],
   dts: true,
   clean: true,
